Started Ts migration for the Client

+++ controllers/interviewController.ts
@@ -6,19 +6,16 @@ import day from "dayjs"
 import { Request, Response } from "express"
 import { NotFoundError, UnauthenticatedError } from "../errors/customError.js"
 import { InterviewMode, InterviewStatus } from "../utils/constants.js"
-type CreateInterviewBody = {
-  company: string
-  position: string
-  interviewAt: Date
-  mode: InterviewMode
-  status: InterviewStatus
-  location?: string
-}
+import {
+  CreateInterviewBody,
+  IdInterviewRequestParams,
+} from "../types/index.js"
 
 export const getAllInterviews = async (
   req: Request,
   res: Response,
 ): Promise<void> => {
+  if (!req.user) throw new UnauthenticatedError("User not authenticated")
   const interviews = await Interview.find({ createdBy: req.user.userId }).sort(
     "-interviewAt",
   )
@@ -29,10 +26,13 @@ export const createInterview = async (
   req: Request,
   res: Response,
 ): Promise<void> => {
+  if (!req.user) throw new UnauthenticatedError("User not authenticated")
   const userId = req.user.userId
   const user = await User.findById(userId)
   if (!user) throw new NotFoundError("User not found")
-  const interview = await Interview.create({ ...req.body, createdBy: userId })
+
+  const body = req.body as CreateInterviewBody
+  const interview = await Interview.create({ ...body, createdBy: userId })
   res.status(StatusCodes.CREATED).json({ interview })
 }
 
@@ -40,7 +40,9 @@ export const getInterviewById = async (
   req: Request,
   res: Response,
 ): Promise<void> => {
-  const { id } = req.params
+  if (!req.user) throw new UnauthenticatedError("User not authenticated")
+  const param = req.params as IdInterviewRequestParams
+  const { id } = param
   const interview = await Interview.findOne({
     _id: id,
     createdBy: req.user.userId,
@@ -51,6 +53,7 @@ export const getInterviewById = async (
 }
 
 export const getUpcomingInterview = async (req: Request, res: Response) => {
+  if (!req.user) throw new UnauthenticatedError("User not authenticated")
   const now = new Date()
   const upcomingCount = await Interview.find({
     createdBy: req.user.userId,
@@ -63,7 +66,9 @@ export const getUpcomingInterview = async (req: Request, res: Response) => {
   res.status(StatusCodes.OK).json({ upcomingCount })
 }
 export const updateInterview = async (req: Request, res: Response) => {
-  const { id } = req.params
+  if (!req.user) throw new UnauthenticatedError("User not authenticated")
+  const param = req.params as IdInterviewRequestParams
+  const { id } = param
   const interview = await Interview.findOneAndUpdate(
     {
       _id: id,
@@ -85,7 +90,9 @@ export const updateInterview = async (req: Request, res: Response) => {
 }
 
 export const deleteInterview = async (req: Request, res: Response) => {
-  const { id } = req.params
+  if (!req.user) throw new UnauthenticatedError("User not authenticated")
+  const param = req.params as IdInterviewRequestParams
+  const { id } = param
 
   const interview = await Interview.findOneAndDelete({
     _id: id,
@@ -101,25 +108,38 @@ export const deleteInterview = async (req: Request, res: Response) => {
   return res.status(StatusCodes.OK).json({ msg: "Interview deleted" })
 }
 
+type monthlyInterviewAgg = {
+  _id: {
+    year: number
+    month: number
+  }
+  count: number
+}
+
+type MonthlyInterview = {
+  date: string
+  count: number
+}
+
 export const showStats = async (req: Request, res: Response): Promise<void> => {
   if (!req.user) throw new UnauthenticatedError("User not  authenticated")
   const now = new Date()
-  let monthlyInterview = await Interview.aggregate([
-    { $match: { createdBy: new mongoose.Types.ObjectId(req.user.userId) } },
-    {
-      $group: {
-        _id: {
-          year: { $year: "$interviewAt" },
-          month: { $month: "$interviewAt" },
+  let monthlyInterview: MonthlyInterview[] = (
+    await Interview.aggregate<monthlyInterviewAgg>([
+      { $match: { createdBy: new mongoose.Types.ObjectId(req.user.userId) } },
+      {
+        $group: {
+          _id: {
+            year: { $year: "$interviewAt" },
+            month: { $month: "$interviewAt" },
+          },
+          count: { $sum: 1 },
         },
-        count: { $sum: 1 },
       },
-    },
-    { $sort: { "_id.year": -1, "_id.month": -1 } },
-    { $limit: 6 },
-  ])
-
-  monthlyInterview = monthlyInterview.map((item) => {
+      { $sort: { "_id.year": -1, "_id.month": -1 } },
+      { $limit: 6 },
+    ])
+  ).map((item: monthlyInterviewAgg) => {
     const {
       _id: { year, month },
       count,
@@ -138,6 +158,7 @@ export const showStats = async (req: Request, res: Response): Promise<void> => {
   })
     .sort("interviewAt")
     .populate("job", "company position jobLocation")
+
   res.status(StatusCodes.OK).json({
     upcomingInterview,
     monthlyInterview,
